requirements.txt
appdirs==1.4.0
backoff==1.3.2
click==6.7
Flask==0.12
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
requests==2.2.0
six==1.10.0
Werkzeug==0.11.15
python-dateutil==2.7.5
